refactor(SFT-497): moving specific integration bundles into their integration directories
