--- packages/plugin/src/Bundles/Integrations/CRM/Pipedrive/PipedriveTokenListener.php
@@ -1,66 +1,0 @@
-<?php
-/**
- * Freeform for Craft CMS.
- *
- * @author        Solspace, Inc.
- * @copyright     Copyright (c) 2008-2022, Solspace, Inc.
- *
- * @see           https://docs.solspace.com/craft/freeform
- *
- * @license       https://docs.solspace.com/license-agreement
- */
-
-namespace Solspace\Freeform\Bundles\Integrations\CRM\Pipedrive;
-
-use Solspace\Freeform\Events\Integrations\OAuth2\InitiateAuthenticationFlowEvent;
-use Solspace\Freeform\Events\Integrations\OAuth2\TokenPayloadEvent;
-use Solspace\Freeform\Integrations\CRM\Pipedrive\PipedriveIntegrationInterface;
-use Solspace\Freeform\Library\Bundles\FeatureBundle;
-use Solspace\Freeform\Library\Exceptions\Integrations\CRMIntegrationNotFoundException;
-use Solspace\Freeform\Library\Integrations\OAuth\OAuth2ConnectorInterface;
-use yii\base\Event;
-
-// TODO: move into integrations and autowire from there
-class PipedriveTokenListener extends FeatureBundle
-{
-    public function __construct()
-    {
-        Event::on(
-            OAuth2ConnectorInterface::class,
-            OAuth2ConnectorInterface::EVENT_INITIATE_AUTHENTICATION_FLOW,
-            [$this, 'onInitAuthentication']
-        );
-
-        Event::on(
-            OAuth2ConnectorInterface::class,
-            OAuth2ConnectorInterface::EVENT_AFTER_AUTHORIZE,
-            [$this, 'onAfterAuthorize']
-        );
-    }
-
-    public function onInitAuthentication(InitiateAuthenticationFlowEvent $event): void
-    {
-        $integration = $event->getIntegration();
-        if (!$integration instanceof PipedriveIntegrationInterface) {
-            return;
-        }
-
-        $event->add('scope', 'base search:read contacts:full deals:full leads:full');
-    }
-
-    public function onAfterAuthorize(TokenPayloadEvent $event): void
-    {
-        $integration = $event->getIntegration();
-        if (!$integration instanceof PipedriveIntegrationInterface) {
-            return;
-        }
-
-        $payload = $event->getResponsePayload();
-
-        if (!isset($payload->api_domain)) {
-            throw new CRMIntegrationNotFoundException("Pipedrive response data doesn't contain the instance API Domain");
-        }
-
-        $integration->setApiDomain($payload->api_domain);
-    }
-}
