--- packages/plugin/src/Bundles/Integrations/MailingLists/ConstantContact/ConstantContactTokenListener.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace Solspace\Freeform\Bundles\Integrations\MailingLists\ConstantContact;
-
-use Solspace\Freeform\Events\Integrations\OAuth2\InitiateAuthenticationFlowEvent;
-use Solspace\Freeform\Integrations\MailingLists\ConstantContact\ConstantContactIntegrationInterface;
-use Solspace\Freeform\Library\Bundles\FeatureBundle;
-use Solspace\Freeform\Library\Integrations\OAuth\OAuth2ConnectorInterface;
-use yii\base\Event;
-
-class ConstantContactTokenListener extends FeatureBundle
-{
-    public function __construct()
-    {
-        Event::on(
-            OAuth2ConnectorInterface::class,
-            OAuth2ConnectorInterface::EVENT_INITIATE_AUTHENTICATION_FLOW,
-            [$this, 'onInitAuthentication']
-        );
-    }
-
-    public function onInitAuthentication(InitiateAuthenticationFlowEvent $event): void
-    {
-        $integration = $event->getIntegration();
-        if (!$integration instanceof ConstantContactIntegrationInterface) {
-            return;
-        }
-
-        $event->add('scope', 'offline_access contact_data account_read account_update campaign_data');
-    }
-}
