--- packages/plugin/src/Bundles/Routing/routes/cp/webhooks.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'freeform/settings/webhooks' => 'freeform/webhooks/index',
-    'freeform/settings/webhooks/new' => 'freeform/webhooks/create',
-    'freeform/settings/webhooks/<id:\d+>' => 'freeform/webhooks/edit',
-];
