+++ packages/plugin/src/Freeform.php
@@ -59,6 +59,7 @@ use Solspace\Freeform\Services\Integrations\ElementsService;
 use Solspace\Freeform\Services\Integrations\IntegrationsService;
 use Solspace\Freeform\Services\Integrations\MailingListsService;
 use Solspace\Freeform\Services\Integrations\PaymentGatewaysService;
+use Solspace\Freeform\Services\Integrations\WebhooksService;
 use Solspace\Freeform\Services\IntegrationsQueueService;
 use Solspace\Freeform\Services\LockService;
 use Solspace\Freeform\Services\LoggerService;
@@ -75,7 +76,6 @@ use Solspace\Freeform\Services\Pro\Payments\StripeService;
 use Solspace\Freeform\Services\Pro\Payments\SubscriptionPlansService;
 use Solspace\Freeform\Services\Pro\Payments\SubscriptionsService;
 use Solspace\Freeform\Services\Pro\RulesService;
-use Solspace\Freeform\Services\Pro\WebhooksService;
 use Solspace\Freeform\Services\Pro\WidgetsService;
 use Solspace\Freeform\Services\RelationsService;
 use Solspace\Freeform\Services\SettingsService;
