+++ packages/plugin/src/Integrations/CRM/HubSpot/EventListeners/HubSpotTokenListener.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Freeform for Craft CMS.
+ *
+ * @author        Solspace, Inc.
+ * @copyright     Copyright (c) 2008-2022, Solspace, Inc.
+ *
+ * @see           https://docs.solspace.com/craft/freeform
+ *
+ * @license       https://docs.solspace.com/license-agreement
+ */
+
+namespace Solspace\Freeform\Integrations\CRM\HubSpot\EventListeners;
+
+use Solspace\Freeform\Events\Integrations\OAuth2\InitiateAuthenticationFlowEvent;
+use Solspace\Freeform\Integrations\CRM\HubSpot\HubSpotIntegrationInterface;
+use Solspace\Freeform\Library\Bundles\FeatureBundle;
+use Solspace\Freeform\Library\Integrations\OAuth\OAuth2ConnectorInterface;
+use yii\base\Event;
+
+class HubSpotTokenListener extends FeatureBundle
+{
+    public function __construct()
+    {
+        Event::on(
+            OAuth2ConnectorInterface::class,
+            OAuth2ConnectorInterface::EVENT_INITIATE_AUTHENTICATION_FLOW,
+            [$this, 'onInitAuthentication']
+        );
+    }
+
+    public function onInitAuthentication(InitiateAuthenticationFlowEvent $event): void
+    {
+        $integration = $event->getIntegration();
+        if (!$integration instanceof HubSpotIntegrationInterface) {
+            return;
+        }
+
+        $event->add('scope', 'crm.objects.contacts.read crm.objects.contacts.write crm.objects.companies.read crm.objects.companies.write crm.objects.deals.read crm.objects.deals.write');
+    }
+}
