+++ packages/plugin/src/Integrations/MailingLists/CampaignMonitor/CampaignMonitorTokenListener.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Solspace\Freeform\Integrations\MailingLists\CampaignMonitor;
+
+use Solspace\Freeform\Events\Integrations\OAuth2\InitiateAuthenticationFlowEvent;
+use Solspace\Freeform\Library\Bundles\FeatureBundle;
+use Solspace\Freeform\Library\Integrations\OAuth\OAuth2ConnectorInterface;
+use yii\base\Event;
+
+class CampaignMonitorTokenListener extends FeatureBundle
+{
+    public function __construct()
+    {
+        Event::on(
+            OAuth2ConnectorInterface::class,
+            OAuth2ConnectorInterface::EVENT_INITIATE_AUTHENTICATION_FLOW,
+            [$this, 'onInitAuthentication']
+        );
+    }
+
+    public function onInitAuthentication(InitiateAuthenticationFlowEvent $event): void
+    {
+        $integration = $event->getIntegration();
+        if (!$integration instanceof CampaignMonitorIntegrationInterface) {
+            return;
+        }
+
+        $event->add('scope', 'ManageLists ImportSubscribers AdministerPersons AdministerAccount');
+    }
+}
