+++ packages/plugin/src/Library/Integrations/Types/Webhooks/WebhookIntegrationInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Solspace\Freeform\Library\Integrations\Types\Webhooks;
+
+use Solspace\Freeform\Library\Integrations\IntegrationInterface;
+
+interface WebhookIntegrationInterface extends IntegrationInterface
+{
+}
